requirements.txt
requests==2.31.0
pandas==2.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
streamlit==1.28.0
plotly==5.15.0
numpy==1.24.3
